pyproject.toml
[project]
name = "noxer-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "apscheduler>=3.11.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-executor>=1.0.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "types-flask-migrate>=4.1.0.20250112",
    "types-requests>=2.32.4.20250611",
]


[tool.flake8]
max-line-length = 79
